wm: Do not set MB theme based on Net/ThemeName
MBWM already supports setting MB theme with "MATCHBOX/THEME",
there is no advantage to also set it with "Net/ThemeName" -- the
expectation that MBWM has a matching theme for all toolkit themes
is not realistic.

Net/ThemeName is used by Gtk+ to set Gtk theme so this change allows
the Gtk theme and MBWM theme to have different names.

Signed-off-by: Jussi Kukkonen <jussi.kukkonen@intel.com>